+++ src/main/java/com/lyhour/developer/phoneshop_studying/exception/ResourceNotFoundOrNot.java
@@ -3,7 +3,7 @@ package com.lyhour.developer.phoneshop_studying.exception;
 import org.springframework.http.HttpStatus;
 
 public class ResourceNotFoundOrNot extends ApiException{
-	public ResourceNotFoundOrNot(String reourceName, Integer id) {
+	public ResourceNotFoundOrNot(String reourceName, Long id) {
 		super(HttpStatus.NOT_FOUND, String.format("%s with id = %d not found", reourceName, id));
 	}
