+++ src/main/java/com/lyhour/developer/phoneshop_studying/service/BrandService.java
@@ -10,8 +10,8 @@ import com.lyhour.developer.phoneshop_studying.entity.Brand;
 @Service
 public interface BrandService {
 	Brand create(Brand brand);
-	Brand getById(Integer id);
-	Brand update(Integer id, Brand brand);
+	Brand getById(Long id);
+	Brand update(Long id, Brand brand);
 	//List<Brand> getBrands(Map<String, String> params);
 	Page<Brand> pagesBrands(Map<String, String> params);
 }
