+++ src/main/java/com/lyhour/developer/phoneshop_studying/service/impl/BrandServiceImpl.java
@@ -27,7 +27,7 @@ public class BrandServiceImpl implements BrandService{
 		return brandRepository.save(brand);
 	}
 	@Override
-	public Brand getById(Integer id) {
+	public Brand getById(Long id) {
 //		Optional<Brand> brandOptional  = brandRepository.findById(id);
 //		if(brandOptional.isPresent()) {
 //			return brandOptional.get();
@@ -42,7 +42,7 @@ public class BrandServiceImpl implements BrandService{
 		
 	}
 	@Override
-	public Brand update(Integer id, Brand brandUpdate) {
+	public Brand update(Long id, Brand brandUpdate) {
 		Brand brand = getById(id);
 		brand.setName(brandUpdate.getName());    // @TODO improve later
 		return brandRepository.save(brand);
